src/requirements.txt
absl-py==1.4.0
array-record==0.4.1
asttokens==2.4.0
astunparse==1.6.3
autoattack==0.1
backcall==0.2.0
cachetools==5.3.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
cmake==3.25.0
comm==0.1.4
contourpy==1.1.1
cycler==0.12.0
debugpy==1.8.0
decorator==5.1.1
dm-tree==0.1.8
etils==1.5.0
exceptiongroup==1.1.3
executing==2.0.0
filelock==3.9.0
flatbuffers==23.5.26
fonttools==4.43.0
fsspec==2023.9.2
gast==0.4.0
google-auth==2.23.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.60.0
grpcio==1.59.0
h5py==3.10.0
idna==3.4
imageio==2.31.4
importlib-resources==6.1.0
ipykernel==6.25.2
ipython==8.16.0
jedi==0.19.0
Jinja2==3.1.2
jupyter_client==8.3.1
jupyter_core==5.3.2
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
lit==15.0.7
Markdown==3.5
MarkupSafe==2.1.2
matplotlib==3.7.2
matplotlib-inline==0.1.6
mizani==0.9.3
ml-dtypes==0.2.0
mpmath==1.3.0
nest-asyncio==1.5.8
networkx==3.1
numpy==1.24.1
oauthlib==3.2.2
opencv-python==4.8.0.76
OpenEXR==1.3.9
opt-einsum==3.3.0
packaging==23.2
pandas==2.0.3
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==3.10.0
plotly==5.15.0
plotnine==0.12.3
promise==2.3
prompt-toolkit==3.0.39
protobuf==3.19.6
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.16.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-version==0.0.2
pytz==2023.3.post1
PyWavelets==1.4.1
pyzmq==25.1.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.21.0
scipy==1.11.1
seaborn==0.12.2
six==1.16.0
stack-data==0.6.3
statsmodels==0.14.0
sympy==1.12
tenacity==8.2.3
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-addons==0.21.0
tensorflow-datasets==4.9.3
tensorflow-estimator==2.10.0
tensorflow-graphics==2021.12.3
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-metadata==1.14.0
termcolor==2.3.0
tf-slim==1.1.0
tifffile==2023.9.26
toml==0.10.2
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
tornado==6.3.3
tqdm==4.65.0
traitlets==5.10.1
trimesh==3.23.5
triton==2.0.0
typeguard==2.13.3
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.13
wcwidth==0.2.8
Werkzeug==3.0.0
wrapt==1.14.1
zipp==3.17.0
